ai-agent/requirements.txt
flask
langgraph==0.1.15
langchain-anthropic==0.1.21
langchain-community==0.2.10
langchain-openai==0.1.19
langchainhub==0.1.20
openai
python-dotenv
azure-core==1.26.4
azure-cosmos==4.3.1
azure-identity==1.17.1
gunicorn